refactor: extract logic from line_webhook.py into usecase.py
